🐛 bug: fix jest configs and reorganize folders
