--- babel.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  "presets": [
-    [
-      "next/babel",
-      {
-        "preset-env": {},
-        "preset-react": {},
-      }
-    ]
-  ],
-  "plugins": []
-}
