+++ components/Stars.js
