+++ components/__tests__/HotelCard.test.js
@@ -2,6 +2,8 @@ import React from 'react';
 import {render, screen} from "@testing-library/react"
 import '@testing-library/jest-dom'
 import HotelCard from "../HotelCard";
+import {expect} from 'jest'
+
 
 
 describe("Competitor Pricing", () => {
