+++ config/jest/cssTransform.js
@@ -1,0 +1,8 @@
+module.exports = {
+  process() {
+    return 'module.exports = {};'
+  },
+  getCacheKey() {
+    return 'cssTransform'
+  },
+}
