+++ utils/mergeHotelsAndPrices.js
@@ -1,6 +1,9 @@
 export default function mergeHotelsAndPrices(hotels, prices) {
   const objectsWithoutPrice = [];
   let mergedResult = []
+  if(!hotels) {
+    hotels = []
+  }
   hotels.map((hotel)=>{
     const pricesObj = prices.find(price => price.id === hotel.id);
     if(!pricesObj) {
